tests/wpt/web-platform-tests/service-workers/service-worker/update-registration-with-type.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Service Worker: Update the registration with a different script type.</title>
<!-- common.js is for guid() -->
<script src="/mixed-content/generic/common.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
// These tests check that a registration is updated correctly with
// different script type. At first Service Worker is register as
// classic script type, then it is re-registered as module script type,
// and vice versa. A main script is also updated at the same time.
promise_test(async t => {
  const key = guid();
  const script = `resources/update-registration-with-type.py?classic_first=1&key=${key}`;
  const scope = 'resources/update-registration-with-type';
  await service_worker_unregister(t, scope);

  // Register with classic script type.
  const firstRegistration = await navigator.serviceWorker.register(script, {
    scope: scope,
    type: 'classic'
  });
  firstRegistration.installing.postMessage(' ');
  let msgEvent = await new Promise(resolve => {
    navigator.serviceWorker.onmessage = resolve;
  });
  assert_equals(msgEvent.data, 'A classic script.');

  // Re-register with module script type.
  const secondRegistration = await navigator.serviceWorker.register(script, {
    scope: scope,
    type: 'module'
  });
  secondRegistration.installing.postMessage(' ');
  msgEvent = await new Promise(resolve => {
    navigator.serviceWorker.onmessage = resolve;
  });
  assert_equals(msgEvent.data, 'A module script.');
  assert_equals(firstRegistration, secondRegistration);
}, 'Update the registration with a different script type (classic => module).');

promise_test(async t => {
  const key = guid();
  const script = `resources/update-registration-with-type.py?classic_first=0&key=${key}`;
  const scope = 'resources/update-registration-with-type';
  await service_worker_unregister(t, scope);

  // Register with module script type.
  const firstRegistration = await navigator.serviceWorker.register(script, {
    scope: scope,
    type: 'module'
  });
  firstRegistration.installing.postMessage(' ');
  let msgEvent = await new Promise(resolve => {
    navigator.serviceWorker.onmessage = resolve;
  });
  assert_equals(msgEvent.data, 'A module script.');

  // Re-register with classic script type.
  const secondRegistration = await navigator.serviceWorker.register(script, {
    scope: scope,
    type: 'classic'
  });
  secondRegistration.installing.postMessage(' ');
  msgEvent = await new Promise(resolve => {
    navigator.serviceWorker.onmessage = resolve;
  });
  assert_equals(msgEvent.data, 'A classic script.');
  assert_equals(firstRegistration, secondRegistration);
}, 'Update the registration with a different script type (module => classic).');
</script>
</body>
